plugins/bootstrap/pyproject.toml
[project]
name = "bootstrap"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = ["pyyaml"]
